Renamed folders and updated namespaces
