+++ api/ExpressedRealms.Powers.Repository/PowerPrerequisites/CreatePrerequisite/CreatePrerequisiteModelValidator.cs
@@ -2,7 +2,7 @@ using ExpressedRealms.Powers.Repository.Powers;
 using FluentValidation;
 using JetBrains.Annotations;
 
-namespace ExpressedRealms.Powers.Repository.PowerPrerequisites.CreatePowerPrerequisite;
+namespace ExpressedRealms.Powers.Repository.PowerPrerequisites.CreatePrerequisite;
 
 [UsedImplicitly]
 internal class CreatePrerequisiteModelValidator : AbstractValidator<CreatePrerequisiteModel>
