+++ api/ExpressedRealms.Powers.Repository/PowerPrerequisites/CreatePrerequisite/ICreatePrerequisiteUseCase.cs
@@ -1,7 +1,7 @@
 using ExpressedRealms.Shared;
 using FluentResults;
 
-namespace ExpressedRealms.Powers.Repository.PowerPrerequisites.CreatePowerPrerequisite;
+namespace ExpressedRealms.Powers.Repository.PowerPrerequisites.CreatePrerequisite;
 
 public interface ICreatePrerequisiteUseCase : IGenericUseCase<Result, CreatePrerequisiteModel>
 {
