+++ api/ExpressedRealms.Powers.Repository/PowerPrerequisites/EditPrerequisite/EditPrerequisiteModel.cs
@@ -1,6 +1,6 @@
-namespace ExpressedRealms.Powers.Repository.PowerPrerequisites.CreatePowerPrerequisite;
+namespace ExpressedRealms.Powers.Repository.PowerPrerequisites.EditPrerequisite;
 
-public class CreatePrerequisiteModel
+public class EditPrerequisiteModel
 {
     public int Id { get; set; }
     public int RequiredAmount { get; set; }
