+++ api/ExpressedRealms.Powers.Repository/PowerPrerequisites/EditPrerequisite/IEditPrerequisiteUseCase.cs
@@ -1,7 +1,7 @@
 using ExpressedRealms.Shared;
 using FluentResults;
 
-namespace ExpressedRealms.Powers.Repository.PowerPrerequisites.EditPowerPrerequisite;
+namespace ExpressedRealms.Powers.Repository.PowerPrerequisites.EditPrerequisite;
 
 public interface IEditPrerequisiteUseCase : IGenericUseCase<Result, EditPrerequisiteModel>
 {
